requirements.txt
apache-airflow[google,redis,postgres]==2.9.0
apache-airflow-providers-google==10.10.0
db-dtypes>=1.4.3
functions-framework>=3.8.3
google-cloud-bigquery>=3.33.0
google-re2==1.0
matplotlib>=3.10.3
nba-api>=1.9.0
numpy>=1.24.0
pandas>=2.2.3
plotly>=6.1.1
psycopg2-binary>=2.9.10
python-dotenv>=1.0.0
pydantic-settings>=2.0a3
pyspark[sql]==3.4.0
requests>=2.31.0
scikit-learn>=1.6.1
seaborn>=0.13.2
sqlalchemy>=1.4.54
torch>=2.7.0
torchvision>=0.22.0